Moved BDN of hash set to separate class
